Move custom fonts loader utility to library source
