+++ src/index.ts
@@ -1,6 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import type { Page } from 'puppeteer-core'
 import twemoji from 'twemoji'
+import path from 'path'
+import fs from 'fs/promises'
 import core from 'puppeteer-core'
 import chrome from 'chrome-aws-lambda'
 
@@ -60,6 +62,19 @@ function emojify(html: string) {
   `
 
   return `<style>${emojiStyle}</style>${emojified}`
+}
+
+export async function fontFilesToBase64(fonts: Array<string>, extension?: string): Promise<Array<string>> {
+  const fontsDirectory = path.resolve('fonts')
+  const fontExtension = extension || 'woff2'
+
+  return await Promise.all(
+    fonts.map(async (font) => {
+      const fontPath = path.join(fontsDirectory, `${font}.${fontExtension}`)
+      const fontFile = await fs.readFile(fontPath)
+      return fontFile.toString('base64')
+    }),
+  )
 }
 
 function pipe(...functions: Function[]): () => Promise<BrowserEnvironment> {
